+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Generated documentation](https://gdx.dotbunny.com/) now including `private` and `internal` classes for visibility.
 - Moved all static `GUIContent` from settings window into new `SettingsContent`.
+- Moved all static layout methods from `SettingsStyles` into `SettingsLayout`.
 - `GDX.Developer.Conditionals` contains constant status indicators of packages used by `GDX`. Useful for determine if a certain featureset is available.
 ### Changed
 - Lowered feature requirement of `com.unity.jobs` to `0.2.7`.